+++ app/admin/(components)/Sidebar.tsx
@@ -21,7 +21,7 @@ import { useSelectedThemeReset } from "@/(shared)/atoms/selectedTheme.atom";
 import { useDrawerState } from "@/(shared)/atoms/drawer.atom";
 import useModal from "@/(shared)/hooks/useModal";
 import HintDialog from "@/(shared)/components/Dialog/Hint/Dialog";
-import { QUERY_KEY } from "@/(shared)/queries/getThemeList";
+import { QUERY_KEY } from "@/admin/apis/theme/getThemeList";
 
 interface Theme {
   id: number;
