+++ app/admin/apis/hint/deleteHint.ts
@@ -4,17 +4,17 @@ import { AxiosError, AxiosResponse } from "axios";
 import { useToastWrite } from "@/(shared)/atoms/toast.atom";
 import { apiClient } from "@/(shared)/lib/reactQueryProvider";
 import { MutationConfigOptions } from "@/(shared)/types";
-import { QUERY_KEY } from "@/(shared)/queries/getThemeList";
+import { QUERY_KEY } from "@/admin/apis/hint/getHintList";
 
 interface Request {
   id: number;
 }
 type Response = void;
 
-const URL_PATH = `/v1/theme`;
+const URL_PATH = `/v1/hint`;
 const MUTATION_KEY = [URL_PATH];
 
-export const deleteTheme = async (req: Request) => {
+export const deleteHint = async (req: Request) => {
   const res = await apiClient.delete<Request, AxiosResponse<Response>>(
     URL_PATH,
     { data: req }
@@ -23,19 +23,19 @@ export const deleteTheme = async (req: Request) => {
   return res.data;
 };
 
-export const useDeleteTheme = (configOptions?: MutationConfigOptions) => {
+export const useDeleteHint = (configOptions?: MutationConfigOptions) => {
   const queryClient = useQueryClient();
   const setToast = useToastWrite();
 
   const info = useMutation<Response, void, Request, void>({
     mutationKey: MUTATION_KEY,
-    mutationFn: (req) => deleteTheme(req),
+    mutationFn: (req) => deleteHint(req),
     ...configOptions?.options,
     onSuccess: () => {
       queryClient.invalidateQueries(QUERY_KEY);
       setToast({
         isOpen: true,
-        title: "테마를 삭제했습니다.",
+        title: "힌트를 삭제했습니다.",
         text: "",
       });
       // console.log("성공 시 실행")
