+++ app/admin/apis/timerImage/deleteTimerImage.ts
@@ -4,38 +4,31 @@ import { AxiosError, AxiosResponse } from "axios";
 import { useToastWrite } from "@/(shared)/atoms/toast.atom";
 import { apiClient } from "@/(shared)/lib/reactQueryProvider";
 import { MutationConfigOptions } from "@/(shared)/types";
-import { QUERY_KEY } from "@/(shared)/queries/getHintList";
+import { QUERY_KEY } from "@/admin/apis/hint/getHintList";
 
-interface Request {
-  id: number;
-}
 type Response = void;
 
-const URL_PATH = `/v1/hint`;
-const MUTATION_KEY = [URL_PATH];
-
-export const deleteHint = async (req: Request) => {
-  const res = await apiClient.delete<Request, AxiosResponse<Response>>(
-    URL_PATH,
-    { data: req }
-  );
+const MUTATION_KEY = ["DeleteTimerImage"];
+const deleteTimerImage = async (themeId: number) => {
+  const URL_PATH = `/v1/theme/timer/${themeId}`;
+  const res = await apiClient.delete<number, AxiosResponse<Response>>(URL_PATH);
 
   return res.data;
 };
 
-export const useDeleteHint = (configOptions?: MutationConfigOptions) => {
+export const useDeleteTimerImage = (configOptions?: MutationConfigOptions) => {
   const queryClient = useQueryClient();
   const setToast = useToastWrite();
 
-  const info = useMutation<Response, void, Request, void>({
+  const info = useMutation<Response, void, number, void>({
     mutationKey: MUTATION_KEY,
-    mutationFn: (req) => deleteHint(req),
+    mutationFn: (req) => deleteTimerImage(req),
     ...configOptions?.options,
     onSuccess: () => {
       queryClient.invalidateQueries(QUERY_KEY);
       setToast({
         isOpen: true,
-        title: "힌트를 삭제했습니다.",
+        title: "타이머 배경을 삭제했습니다.",
         text: "",
       });
       // console.log("성공 시 실행")
